requirements.txt
Django==3.2.3
djangorestframework==3.12.4
mysqlclient==2.0.3
django-mysql==3.11.1
django-cors-headers==3.7.0
pyJWT==2.0.1
Faker
